+++ test/dummy/config/routes.rb
@@ -5,7 +5,19 @@ Rails.application.routes.draw do
 
   mount Folio::Engine => '/'
 
-  scope '/:locale', locale: /#{I18n.available_locales.join('|')}/ do
-    get '/*path', to: 'pages#show', as: 'page'
+  if Rails.application.config.folio_pages_translations
+    scope '/:locale', locale: /#{I18n.available_locales.join('|')}/ do
+      if Rails.application.config.folio_pages_ancestry
+        get '/*path', to: 'pages#show', as: 'page'
+      else
+        resources :pages, only: [:show], path: ''
+      end
+    end
+  else
+    if Rails.application.config.folio_pages_ancestry
+      get '/*path', to: 'pages#show', as: 'page'
+    else
+      resources :pages, only: [:show], path: ''
+    end
   end
 end